+++ src/Connection/AbstractPdoConnection.php
@@ -33,7 +33,17 @@ abstract class AbstractPdoConnection implements ConnectionInterface
 
     abstract protected function createDriver(): DriverInterface;
 
-    protected function __construct(
+    public static function create(
+        string $dsn,
+        string $username,
+        string $password,
+        array $options,
+        bool $isLogging
+    ): ConnectionInterface {
+        return new static($dsn, $username, $password, $options, $isLogging);
+    }
+
+    public function __construct(
         string $dsn,
         string $username,
         string $password,