+++ src/Connection/MySqlPdoConnection.php
@@ -6,7 +6,6 @@ namespace WoohooLabs\Larva\Connection;
 use PDO;
 use WoohooLabs\Larva\Driver\Driver;
 use WoohooLabs\Larva\Driver\DriverInterface;
-use WoohooLabs\Larva\Driver\Mysql\MySqlConditionsTranslator;
 use WoohooLabs\Larva\Driver\Mysql\MySqlDeleteTranslator;
 use WoohooLabs\Larva\Driver\Mysql\MySqlInsertTranslator;
 use WoohooLabs\Larva\Driver\Mysql\MySqlSelectTranslator;
@@ -14,7 +13,7 @@ use WoohooLabs\Larva\Driver\Mysql\MySqlUpdateTranslator;
 
 class MySqlPdoConnection extends AbstractPdoConnection
 {
-    public static function create(
+    public function __construct(
         string $host,
         int $port,
         string $database,
@@ -25,11 +24,11 @@ class MySqlPdoConnection extends AbstractPdoConnection
         array $modes = [],
         array $options = [],
         bool $isLogging = false
-    ): MySqlPdoConnection {
+    ) {
         $dsn = "mysql:host=$host;dbname=$database;port=$port;charset=$charset";
         $options[PDO::MYSQL_ATTR_INIT_COMMAND] = self::getInitCommand($charset, $collation, $modes);
 
-        return new MySqlPdoConnection($dsn, $username, $password, $options, $isLogging);
+        parent::__construct($dsn, $username, $password, $options, $isLogging);
     }
 
     protected function createDriver(): DriverInterface