+++ src/Driver/TranslatedQuerySegment.php
@@ -13,18 +13,18 @@ class TranslatedQuerySegment
     /**
      * @var array
      */
-    private $params;
+    private $params = [];
 
     public function __construct($sql = "", array $params = [])
     {
         $this->sql = $sql;
-        $this->params = $params;
+        $this->addParams($params);
     }
 
     public function add(string $sql, array $params = [])
     {
         $this->sql .= $sql;
-        $this->params = array_merge($this->params, $params);
+        $this->addParams($params);
     }
 
     public function getSql(): string
@@ -35,5 +35,16 @@ class TranslatedQuerySegment
     public function getParams(): array
     {
         return $this->params;
+    }
+
+    private function addParams(array $params)
+    {
+        foreach ($params as $param) {
+            if (is_bool($param)) {
+                $param = (int) $param;
+            }
+
+            $this->params[] = $param;
+        }
     }
 }