+++ CHANGELOG.md
@@ -16,7 +16,8 @@ FIXED:
 - `Bool` values are now casted to `int` when using prepared statements
 - Return type of `SelectQueryBuilder::fetchColumn()` became `mixed` so that it can now be used to also retrieve numbers
 - Adding condition groups via `ConditionBuilder::addConditionGroup()` when originally there was no condition
-- Translation of subselect statements
+- Subselect statements are now translated properly when using the MySQL driver
+- Columns in the `SELECT` clause are now translated properly when using the MySQL driver
 
 ## 0.3.1 - 2017-02-24